Ported to typescript
